Raykov's factor reliability coefficient

This function computes Raykov's reliability coefficient (RRC) for factors from confirmatory factor analyses, a measure which is commonly seen as a more accurate one than that of Cronbach's alpha which is computed based on the assumption of tau-equivalent measures. It computes reliability coefficients for factors with and without correlated errors.

relicoef(mod)

Examples

library(lavaan)
#> This is lavaan 0.6-9
#> lavaan is FREE software! Please report any bugs.
model.01 <- '
           Col =~ respected + secure
           Ind =~ accomplish + self_fulfil + self_respect
           '
est.model.01 <- cfa(model.01, data=values)
relicoef(est.model.01)
#>   Latent       RRC
#> 1    Col 0.7696239
#> 2    Ind 0.8311486