Single-trial responses from a Stroop task (85 subjs).

A dataset containing individual responses (RT/accuracy) from 85 individuals completing a standard Stroop-task. Raw data and description available at https://osf.io/kxpqu/.

stroop

Format

A data frame with 49309 rows and 9 variables:

subj

subject ID

trial

trial number in experiment

color

color of the presented word

condition

one of "neutral" (the presented word was XXXXX), "congruent" (the meaning of the presented word was identical to the color of the word) and "incongruent" (the meaning of the word was a different color than the print color of the stimulus)

correct

whether the response was correct

response

the color indicated by the participant (by button press)

RT

reaction time (in sec) between presentation of the stimulus and the response

age

participants age at time of data-acquisition

female

indicator variable (1=female, 0=male)

Source

https://osf.io/kxpqu/