75 mood items from the Motivational State Questionnaire for 3896 participants

Emotions may be described either as discrete emotions or in dimensional terms. The Motivational State Questionnaire (MSQ) was developed to study emotions in laboratory and field settings. The data can be well described in terms of a two dimensional solution of energy vs tiredness and tension versus calmness. Additional items include what time of day the data were collected and a few personality questionnaire scores.

data(msq)

Format

A data frame with 3896 observations on the following 92 variables.

MSQ_Time

Time of day the data were collected

active

a numeric vector

afraid

a numeric vector

alert

a numeric vector

alone

a numeric vector

angry

a numeric vector

aroused

a numeric vector

ashamed

a numeric vector

astonished

a numeric vector

at-ease

a numeric vector

at-rest

a numeric vector

attentive

a numeric vector

blue

a numeric vector

bored

a numeric vector

calm

a numeric vector

clutched-up

a numeric vector

confident

a numeric vector

content

a numeric vector

delighted

a numeric vector

depressed

a numeric vector

determined

a numeric vector

distressed

a numeric vector

drowsy

a numeric vector

dull

a numeric vector

elated

a numeric vector

energetic

a numeric vector

enthusiastic

a numeric vector

excited

a numeric vector

fearful

a numeric vector

frustrated

a numeric vector

full-of-pep

a numeric vector

gloomy

a numeric vector

grouchy

a numeric vector

guilty

a numeric vector

happy

a numeric vector

hostile

a numeric vector

inspired

a numeric vector

intense

a numeric vector

interested

a numeric vector

irritable

a numeric vector

jittery

a numeric vector

kindly

a numeric vector

lively

a numeric vector

lonely

a numeric vector

nervous

a numeric vector

placid

a numeric vector

pleased

a numeric vector

proud

a numeric vector

quiescent

a numeric vector

quiet

a numeric vector

relaxed

a numeric vector

sad

a numeric vector

satisfied

a numeric vector

scared

a numeric vector

scornful

a numeric vector

serene

a numeric vector

sleepy

a numeric vector

sluggish

a numeric vector

sociable

a numeric vector

sorry

a numeric vector

still

a numeric vector

strong

a numeric vector

surprised

a numeric vector

tense

a numeric vector

tired

a numeric vector

unhappy

a numeric vector

upset

a numeric vector

vigorous

a numeric vector

wakeful

a numeric vector

warmhearted

a numeric vector

wide-awake

a numeric vector

anxious

a numeric vector

idle

a numeric vector

cheerful

a numeric vector

inactive

a numeric vector

tranquil

a numeric vector

EA

Thayer's Energetic Arousal Scale

TA

Thayer's Tense Arousal Scale

PA

Positive Affect scale

NegAff

Negative Affect scale

Extraversion

Extraversion from the Eysenck Personality Inventory

Neuroticism

Neuroticism from the Eysenck Personality Inventory

Lie

Lie from the EPI

Sociability

The sociability subset of the Extraversion Scale

Impulsivity

The impulsivity subset of the Extraversions Scale

MSQ_Round

Rounded time of day

scale

a factor with levels msq r original or revised msq

ID

subject ID

exper

Which study were the data collected: a factor with levels AGES BING BORN CART CITY COPE EMIT FAST Fern FILM FLAT Gray imps item knob MAPS mite pat-1 pat-2 PATS post RAFT Rim.1 Rim.2 rob-1 rob-2 ROG1 ROG2 SALT sam-1 sam-2 SAVE/PATS sett swam swam-2 TIME VALE-1 VALE-2 VIEW

condition

What was the experimental condition after the msq was given

TOD

a numeric vector

TOD24

a numeric vector

Details

The Motivational States Questionnaire (MSQ) is composed of 72 items, which represent the full affective range (Revelle & Anderson, 1996). The MSQ consists of 20 items taken from the Activation-Deactivation Adjective Check List (Thayer, 1986), 18 from the Positive and Negative Affect Schedule (PANAS, Watson, Clark, & Tellegen, 1988) along with the items used by Larsen and Diener (1992). The response format was a four-point scale that corresponds to Russell and Carroll's (1999) "ambiguous--likely-unipolar format" and that asks the respondents to indicate their current standing (``at this moment") with the following rating scale:
0----------------1----------------2----------------3
Not at all A little Moderately Very much
The original version of the MSQ included 70 items. Intermediate analyses (done with 1840 subjects) demonstrated a concentration of items in some sections of the two dimensional space, and a paucity of items in others. To begin correcting this, 3 items from redundantly measured sections (alone, kindly, scornful) were removed, and 5 new ones (anxious, cheerful, idle, inactive, and tranquil) were added. Thus, the correlation matrix is missing the correlations between items 5, 42, and 55 and 72-76.

Procedure. The data were collected over nine years, as part of a series of studies examining the effects of personality and situational factors on motivational state and subsequent cognitive performance. In each of 38 studies, prior to any manipulation of motivational state, participants signed a consent form and filled out the MSQ. (The procedures of the individual studies are irrelevant to this data set and could not affect the responses to the MSQ, since this instrument was completed before any further instructions or tasks).

In addition to the MSQ, there are 5 scales from the Eysenck Personality Inventory.

Source

Data collecte at the Personality, Motivation, and Cognition Laboratory, Northwestern University.

References

William Revelle and Kristen Joan Anderson (1997) Personality, motivation and cognitive performance: Final report to the Army Research Institute on contract MDA 903-93-K-0008

Rafaeli, Eshkol and Revelle, William (2006), A premature consensus: Are happiness and sadness truly opposite affects? Motivation and Emotion, 30, 1, 1-12.

Examples

data(msq)
summary(msq)
#>     MSQ_Time         active          afraid           alert      
#>  Min.   : 5.75   Min.   :0.000   Min.   :0.0000   Min.   :0.000  
#>  1st Qu.: 9.30   1st Qu.:0.000   1st Qu.:0.0000   1st Qu.:0.000  
#>  Median :13.23   Median :1.000   Median :0.0000   Median :1.000  
#>  Mean   :13.49   Mean   :1.028   Mean   :0.1157   Mean   :1.152  
#>  3rd Qu.:16.03   3rd Qu.:2.000   3rd Qu.:0.0000   3rd Qu.:2.000  
#>  Max.   :22.75   Max.   :3.000   Max.   :3.0000   Max.   :3.000  
#>  NA's   :1       NA's   :6       NA's   :5        NA's   :11     
#>      alone            angry           aroused        ashamed      
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.00   Min.   :0.0000  
#>  1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.00   1st Qu.:0.0000  
#>  Median :0.0000   Median :0.0000   Median :0.00   Median :0.0000  
#>  Mean   :0.5354   Mean   :0.2151   Mean   :0.71   Mean   :0.1199  
#>  3rd Qu.:1.0000   3rd Qu.:0.0000   3rd Qu.:1.00   3rd Qu.:0.0000  
#>  Max.   :3.0000   Max.   :3.0000   Max.   :3.00   Max.   :3.0000  
#>  NA's   :2058     NA's   :9        NA's   :6      NA's   :11      
#>    astonished       at-ease         at-rest        attentive    
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:0.000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:1.000  
#>  Median :0.000   Median :2.000   Median :1.000   Median :1.000  
#>  Mean   :0.127   Mean   :1.588   Mean   :1.202   Mean   :1.179  
#>  3rd Qu.:0.000   3rd Qu.:2.000   3rd Qu.:2.000   3rd Qu.:2.000  
#>  Max.   :3.000   Max.   :3.000   Max.   :3.000   Max.   :3.000  
#>  NA's   :13      NA's   :17      NA's   :17      NA's   :6      
#>       blue            bored           calm        clutched-up    
#>  Min.   :0.0000   Min.   :0.00   Min.   :0.000   Min.   :0.0000  
#>  1st Qu.:0.0000   1st Qu.:1.00   1st Qu.:1.000   1st Qu.:0.0000  
#>  Median :0.0000   Median :1.00   Median :2.000   Median :0.0000  
#>  Mean   :0.3739   Mean   :1.31   Mean   :1.549   Mean   :0.3726  
#>  3rd Qu.:1.0000   3rd Qu.:2.00   3rd Qu.:2.000   3rd Qu.:1.0000  
#>  Max.   :3.0000   Max.   :3.00   Max.   :3.000   Max.   :3.0000  
#>  NA's   :5        NA's   :4      NA's   :82      NA's   :23      
#>    confident        content        delighted        depressed     
#>  Min.   :0.000   Min.   :0.000   Min.   :0.0000   Min.   :0.0000  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:0.0000   1st Qu.:0.0000  
#>  Median :2.000   Median :1.000   Median :1.0000   Median :0.0000  
#>  Mean   :1.491   Mean   :1.447   Mean   :0.7751   Mean   :0.4107  
#>  3rd Qu.:2.000   3rd Qu.:2.000   3rd Qu.:1.0000   3rd Qu.:1.0000  
#>  Max.   :3.000   Max.   :3.000   Max.   :3.0000   Max.   :3.0000  
#>  NA's   :7       NA's   :22      NA's   :6        NA's   :17      
#>    determined      distressed         drowsy           dull       
#>  Min.   :0.000   Min.   :0.0000   Min.   :0.000   Min.   :0.0000  
#>  1st Qu.:0.000   1st Qu.:0.0000   1st Qu.:0.000   1st Qu.:0.0000  
#>  Median :1.000   Median :0.0000   Median :1.000   Median :1.0000  
#>  Mean   :1.041   Mean   :0.3963   Mean   :1.162   Mean   :0.7597  
#>  3rd Qu.:2.000   3rd Qu.:1.0000   3rd Qu.:2.000   3rd Qu.:1.0000  
#>  Max.   :3.000   Max.   :3.0000   Max.   :3.000   Max.   :3.0000  
#>  NA's   :7       NA's   :8        NA's   :12      NA's   :9       
#>      elated         energetic       enthusiastic       excited      
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.0000   Min.   :0.0000  
#>  1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.0000  
#>  Median :0.0000   Median :1.0000   Median :1.0000   Median :0.0000  
#>  Mean   :0.5246   Mean   :0.8416   Mean   :0.9203   Mean   :0.7159  
#>  3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:2.0000   3rd Qu.:1.0000  
#>  Max.   :3.0000   Max.   :3.0000   Max.   :3.0000   Max.   :3.0000  
#>  NA's   :15       NA's   :6        NA's   :6        NA's   :6       
#>     fearful         frustrated      full-of-pep         gloomy     
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.0000   Min.   :0.000  
#>  1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.000  
#>  Median :0.0000   Median :0.0000   Median :1.0000   Median :0.000  
#>  Mean   :0.1324   Mean   :0.4901   Mean   :0.8082   Mean   :0.398  
#>  3rd Qu.:0.0000   3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:1.000  
#>  Max.   :3.0000   Max.   :3.0000   Max.   :3.0000   Max.   :3.000  
#>  NA's   :20       NA's   :11       NA's   :12       NA's   :12     
#>     grouchy           guilty           happy          hostile      
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.000   Min.   :0.0000  
#>  1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.000   1st Qu.:0.0000  
#>  Median :0.0000   Median :0.0000   Median :1.000   Median :0.0000  
#>  Mean   :0.4724   Mean   :0.1817   Mean   :1.225   Mean   :0.2919  
#>  3rd Qu.:1.0000   3rd Qu.:0.0000   3rd Qu.:2.000   3rd Qu.:0.0000  
#>  Max.   :3.0000   Max.   :3.0000   Max.   :3.000   Max.   :3.0000  
#>  NA's   :5        NA's   :5        NA's   :16      NA's   :11      
#>     inspired         intense         interested      irritable    
#>  Min.   :0.0000   Min.   :0.0000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.000   1st Qu.:0.000  
#>  Median :0.0000   Median :0.0000   Median :1.000   Median :0.000  
#>  Mean   :0.5879   Mean   :0.6868   Mean   :1.144   Mean   :0.551  
#>  3rd Qu.:1.0000   3rd Qu.:1.0000   3rd Qu.:2.000   3rd Qu.:1.000  
#>  Max.   :3.0000   Max.   :3.0000   Max.   :3.000   Max.   :3.000  
#>  NA's   :6        NA's   :7        NA's   :12      NA's   :16     
#>     jittery           kindly          lively           lonely      
#>  Min.   :0.0000   Min.   :0.000   Min.   :0.0000   Min.   :0.0000  
#>  1st Qu.:0.0000   1st Qu.:0.000   1st Qu.:0.0000   1st Qu.:0.0000  
#>  Median :0.0000   Median :1.000   Median :1.0000   Median :0.0000  
#>  Mean   :0.5925   Mean   :1.184   Mean   :0.9205   Mean   :0.4702  
#>  3rd Qu.:1.0000   3rd Qu.:2.000   3rd Qu.:2.0000   3rd Qu.:1.0000  
#>  Max.   :3.0000   Max.   :3.000   Max.   :3.0000   Max.   :3.0000  
#>  NA's   :6        NA's   :2060    NA's   :10       NA's   :6       
#>     nervous           placid         pleased           proud       
#>  Min.   :0.0000   Min.   :0.000   Min.   :0.0000   Min.   :0.0000  
#>  1st Qu.:0.0000   1st Qu.:1.000   1st Qu.:0.0000   1st Qu.:0.0000  
#>  Median :0.0000   Median :1.000   Median :1.0000   Median :1.0000  
#>  Mean   :0.3537   Mean   :1.198   Mean   :0.9009   Mean   :0.9427  
#>  3rd Qu.:1.0000   3rd Qu.:2.000   3rd Qu.:1.0000   3rd Qu.:2.0000  
#>  Max.   :3.0000   Max.   :3.000   Max.   :3.0000   Max.   :3.0000  
#>  NA's   :17       NA's   :19      NA's   :13       NA's   :7       
#>    quiescent          quiet          relaxed           sad        
#>  Min.   :0.0000   Min.   :0.000   Min.   :0.000   Min.   :0.0000  
#>  1st Qu.:0.0000   1st Qu.:1.000   1st Qu.:1.000   1st Qu.:0.0000  
#>  Median :1.0000   Median :1.000   Median :2.000   Median :0.0000  
#>  Mean   :0.6519   Mean   :1.394   Mean   :1.677   Mean   :0.3438  
#>  3rd Qu.:1.0000   3rd Qu.:2.000   3rd Qu.:2.000   3rd Qu.:1.0000  
#>  Max.   :3.0000   Max.   :3.000   Max.   :3.000   Max.   :3.0000  
#>  NA's   :136      NA's   :5       NA's   :7       NA's   :10      
#>    satisfied         scared          scornful          serene    
#>  Min.   :0.000   Min.   :0.0000   Min.   :0.0000   Min.   :0.00  
#>  1st Qu.:1.000   1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:1.00  
#>  Median :1.000   Median :0.0000   Median :0.0000   Median :1.00  
#>  Mean   :1.344   Mean   :0.1668   Mean   :0.1687   Mean   :1.18  
#>  3rd Qu.:2.000   3rd Qu.:0.0000   3rd Qu.:0.0000   3rd Qu.:2.00  
#>  Max.   :3.000   Max.   :3.0000   Max.   :3.0000   Max.   :3.00  
#>  NA's   :7       NA's   :10       NA's   :2058     NA's   :12    
#>      sleepy         sluggish        sociable         sorry       
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.0000  
#>  1st Qu.:0.000   1st Qu.:0.000   1st Qu.:1.000   1st Qu.:0.0000  
#>  Median :1.000   Median :1.000   Median :1.000   Median :0.0000  
#>  Mean   :1.246   Mean   :1.171   Mean   :1.336   Mean   :0.2015  
#>  3rd Qu.:2.000   3rd Qu.:2.000   3rd Qu.:2.000   3rd Qu.:0.0000  
#>  Max.   :3.000   Max.   :3.000   Max.   :3.000   Max.   :3.0000  
#>  NA's   :16      NA's   :8       NA's   :6       NA's   :15      
#>      still           strong        surprised         tense       
#>  Min.   :0.000   Min.   :0.000   Min.   :0.000   Min.   :0.0000  
#>  1st Qu.:1.000   1st Qu.:0.000   1st Qu.:0.000   1st Qu.:0.0000  
#>  Median :1.000   Median :1.000   Median :0.000   Median :0.0000  
#>  Mean   :1.182   Mean   :1.048   Mean   :0.155   Mean   :0.5219  
#>  3rd Qu.:2.000   3rd Qu.:2.000   3rd Qu.:0.000   3rd Qu.:1.0000  
#>  Max.   :3.000   Max.   :3.000   Max.   :3.000   Max.   :3.0000  
#>  NA's   :12      NA's   :7       NA's   :6       NA's   :10      
#>      tired          unhappy           upset           vigorous     
#>  Min.   :0.000   Min.   :0.0000   Min.   :0.0000   Min.   :0.0000  
#>  1st Qu.:1.000   1st Qu.:0.0000   1st Qu.:0.0000   1st Qu.:0.0000  
#>  Median :1.000   Median :0.0000   Median :0.0000   Median :0.0000  
#>  Mean   :1.386   Mean   :0.3727   Mean   :0.3086   Mean   :0.6397  
#>  3rd Qu.:2.000   3rd Qu.:1.0000   3rd Qu.:0.0000   3rd Qu.:1.0000  
#>  Max.   :3.000   Max.   :3.0000   Max.   :3.0000   Max.   :3.0000  
#>  NA's   :10      NA's   :5        NA's   :8        NA's   :10      
#>     wakeful       warmhearted      wide-awake        anxious      
#>  Min.   :0.000   Min.   :0.000   Min.   :0.0000   Min.   :0.0000  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:0.0000   1st Qu.:0.0000  
#>  Median :1.000   Median :1.000   Median :1.0000   Median :0.0000  
#>  Mean   :1.186   Mean   :1.453   Mean   :0.9405   Mean   :0.6717  
#>  3rd Qu.:2.000   3rd Qu.:2.000   3rd Qu.:2.0000   3rd Qu.:1.0000  
#>  Max.   :3.000   Max.   :3.000   Max.   :3.0000   Max.   :3.0000  
#>  NA's   :10      NA's   :7       NA's   :12       NA's   :1849    
#>       idle           cheerful        inactive        tranquil    
#>  Min.   :0.0000   Min.   :0.000   Min.   :0.000   Min.   :0.000  
#>  1st Qu.:0.0000   1st Qu.:0.000   1st Qu.:0.000   1st Qu.:1.000  
#>  Median :1.0000   Median :1.000   Median :1.000   Median :1.000  
#>  Mean   :0.9561   Mean   :0.977   Mean   :1.078   Mean   :1.287  
#>  3rd Qu.:2.0000   3rd Qu.:2.000   3rd Qu.:2.000   3rd Qu.:2.000  
#>  Max.   :9.0000   Max.   :3.000   Max.   :3.000   Max.   :3.000  
#>  NA's   :1846     NA's   :1850    NA's   :1846    NA's   :1843   
#>        EA              TA              PA             NegAff      
#>  Min.   : 0.00   Min.   : 0.00   Min.   : 0.000   Min.   : 0.000  
#>  1st Qu.: 6.00   1st Qu.: 8.00   1st Qu.: 4.000   1st Qu.: 0.000  
#>  Median :11.00   Median :11.00   Median : 9.000   Median : 2.000  
#>  Mean   :11.74   Mean   :11.34   Mean   : 9.735   Mean   : 2.982  
#>  3rd Qu.:17.00   3rd Qu.:14.00   3rd Qu.:14.000   3rd Qu.: 4.000  
#>  Max.   :30.00   Max.   :33.00   Max.   :30.000   Max.   :28.000  
#>  NA's   :598     NA's   :641     NA's   :572      NA's   :592     
#>   Extraversion    Neuroticism         Lie         Sociability    
#>  Min.   : 0.00   Min.   : 0.00   Min.   :0.000   Min.   : 0.000  
#>  1st Qu.:11.00   1st Qu.: 7.00   1st Qu.:1.000   1st Qu.: 6.000  
#>  Median :14.00   Median :10.00   Median :2.000   Median : 8.000  
#>  Mean   :13.55   Mean   :10.39   Mean   :2.113   Mean   : 7.717  
#>  3rd Qu.:16.68   3rd Qu.:14.00   3rd Qu.:3.000   3rd Qu.:10.000  
#>  Max.   :24.00   Max.   :24.00   Max.   :9.000   Max.   :13.000  
#>  NA's   :6       NA's   :6       NA's   :6       NA's   :6       
#>   Impulsivity      MSQ_Round     scale            ID             exper     
#>  Min.   :0.000   Min.   : 6.00   msq:1840   Min.   :   1.0   Rim.1  : 342  
#>  1st Qu.:3.000   1st Qu.: 9.00   r  :2056   1st Qu.:  30.0   Rim.2  : 342  
#>  Median :4.000   Median :13.00              Median :  67.0   sam-1  : 324  
#>  Mean   :4.449   Mean   :13.44              Mean   : 159.4   sam-2  : 324  
#>  3rd Qu.:6.000   3rd Qu.:16.00              3rd Qu.: 140.0   FLAT   : 170  
#>  Max.   :9.000   Max.   :23.00              Max.   :4184.0   MAPS   : 160  
#>  NA's   :6       NA's   :1099                                (Other):2234  
#>    condition          TOD            TOD24      
#>  Min.   :1.000   Min.   : 5.00   Min.   : 5.00  
#>  1st Qu.:1.000   1st Qu.: 9.00   1st Qu.: 9.00  
#>  Median :2.000   Median :13.00   Median :15.00  
#>  Mean   :2.568   Mean   :13.19   Mean   :18.13  
#>  3rd Qu.:4.000   3rd Qu.:16.00   3rd Qu.:19.50  
#>  Max.   :5.000   Max.   :22.50   Max.   :46.50  
#>                  NA's   :1       NA's   :1102